Remove v from auto github url
Previously, when setting the version for a github url, the build-pkg
script would automatically add "v" in front when searching for its tag.
However, not all projects follow this standard, so this has been removed
from build-pkg and the packages that used this feature have been updated
with an explicit "v" in their version values.