Add further support to configure builds
This updates the host target to linux-gnu. Even though this will be a
musl build, linux-musl is not recognised by all configure scripts, and
setting it as gnu tricks it into playing ball. As CC is explictly set,
this is not a problem.

We also add flags to allow the make targets to be changed, as with
"make" style build jobs.