Update directory values
Previously, we set the prefix value to our target folder, however this
was causing this to be erroneously backed into some of the built
binaries. Instead we "correctly" set the PREFIX to be the root, and set
the DESTDIR to our target on our host machine.

This required a few extra tweaks to ensure all projects built as
expected.