Add DHCP-CNI system pod
This pod is an extension of the host DHCP pod, however it extends it by
bridging the host's network interface with veth pairs given to each
container. This allows all CNI containers to sit on the host subnet,
with an IP address assigned to them by the network's DHCP.

This means that containers have direct L2 access to the physical network
(and are granted their own mac addresses). If you were to give a pod
NET_RAW capabilities, it can do its own ARP or other low level protocols
if desired, similarly with NET_ADMIN, it can change, or give itself
extra, IP addresses. (WARNING: this can be abused! Do not give those
capabilities lightly!)

Currently, there is no "smooth" handover between host dhcp and this
dhcp-cni pod (due to the requirement that host routes and address being
moved over to the bridge) so this needs to be installed _instead_ of the
dhcp sys pod as a static manifest.

This may be something that is worked on in future releases - a hand over
is not impossible (as it was achieved during testing), but it adds
complexity as there is a certain amount of coupling between the host
interface and container interfaces with this setup.