Add base SSL certificates
Required by the system to just any TLS site (eg for image pulls!). Until
this point, we loaded these in via the etc/ssl folder on the SD boot
drive which bind mounts to /etc/ssl/certs to allow admins to inject
their own certs. However, it makes sense for the initramfs to come
prepackaged with the "standard" trusted certs.