Restructure partitions
We are now considering the boot partition as read only. This is for a
few reasons:

1) For security and stability - having this partition mounted ro by the
   host operating system makes it harder to inadvertently give a
   container write access: even if you mounted / as a volume, your pods
   can't accidentally or maliciously bork the boot code.

2) Having a "boot is ro" attitude allows the system to support genuinely
   read only devices (eg Raspberry Pis with a flashed EPROM boot drive).
   KiOS can now safely be installed onto any such system without
   special changes.

NB: There is no requirement to use two partitions if you don't like this
design. If you set the bootpart= and datapart= kernel cmdline flags to
the same device, KiOS will not complain at this and that is a completely
safe configuration.

If you do use separate partitions and would like to override this RO
limit, there are two methods:

1) Stop the device and manually edit the partition on another machine
   with an SD card reader.

2) Run a privileged pod with a BiDirectional mount to the rootfs, and
   have it explicitly mount the original drive again with rw
   permissions. (NB: This is intentionally kind of possible but also a
   faff to strike the balance between letting people do what they want
   with their own setup, whilst responsibly discouraging something that
   could potentially be problematic).