Support proper system shutdown
This extends the Graceful Node Shutdown shim to signal the init process
when it is happy that the kubelet shutdown process is complete.

init was already designed to process a shutdown on this signal, but we
have updated the inittab to also ensure that a final umount -a -r is
called.