Remove kubeadm flags file support
The flags that this file sets are kind of dumb. We also don't really
want to allow these flags to be updated as the system cannot support
anything other than the crio runtime (giving people the option to change
this just gives them option to brick their box).

Everything else (as far as I am aware) can be configured via the kubelet
config file, which we do allow users to edit.