Fix kubelet init branch
There is an if condition at the start of the kubelet init script which
starts the kubelet in standalone mode if it detects that its config is
missing; this is so it can bootstrap itself. This had not been tested
and had some typos in it. This has been fixed and this now works.