Add an init pod
We have added a new kind of static pod: init.yaml. If required, its
manifest should be in the root of the boot partition. kiOS will copy
this file into the kubelet static manifests directory _only_ if it
detects the node has not been bootstraped.

This can be used for "one time setup" tasks which you really don't want
to restart each time the node starts up. If you do want it to run every
time (which may sometimes be required, just use a standard system static
pod).

One time is nice for kubeadm so that we can use it to setup sensible
defaults when first running (eg kube-proxy and coredns) but if you
change those without worrying about the system ignoring you.