Remove init.yaml on non-first boots
The special init.yaml file is loaded iff the node is booting and detects
that it does not have the required config. However, previously it was
never unloaded meaning it would actually end up running on every boot.

This adds a special check for "normal" boots which explictly removes the
init.yaml manifest to ensure that it is not present.

NB: It should be noted that you will NOT see logs for the "old" version
of this pod.