Move .config out of build-pkg
Previously, the build-pkg script would automatically copy a "config"
file into the src directory for make jobs if detected. However, as
busybox needs to be able to do its own find and replace on this file, it
makes sense to have this done manually by pre_build - this only affects
the linux and busybox packages.