Switch to usleep for wait_for_file
During the boot process, the kubelet init script needs to wait for files
to exist before proceeding with some steps. Currently it does this with
a sleepy while loop (while ! file; do sleep; done). This was sleeping
for 1 second per check - we've now dropped sleep from busybox and
instead switched to usleep, and set wait_for_file to sleep for just 50
microseconds - 1/20th of a second.

At some point, I am hoping to switch to inotify so that we can a) have
instant notifications and b) not have to do any form of polling at all.