Add wait-for-time container
This tiny container is designed to be used as an init container for pods
with main runtimes that expect the system time to be correctish. The
container will not exit until the system time is "sensible". Kubeadm's
manifest has been updated to use this to ensure it does not generate
certificates in 1970.

NB: This means that kubeadm can now be pre-primed for faster startup.

Fixes: #1