Update to cri-o 1.24.2
The [big change][pr-5811] in 1.24 is that the requirement to have
conntrack installed has been removed, which allows us to delete _a lot_
of annoying-to-compile libraries.

NB: This version upgrade also changes the default pause container from
`k8s.gcr.io/pause:3.6` to `registry.k8s.io/pause:3.6`. You need to be
aware of this when priming system containers.

[pr-5811]: https://github.com/cri-o/cri-o/pull/5811