Fix setting of hostname in bootstrap container
I had originally assumed that the bootstrap container, as it runs with
hostNetwork=true set, would share the host's UTS namespace - however it
turns out that this is not the case. All containers always get their own
dedicated UTS namespace.

So instead init now supports reading /etc/hostname after kubelet has
been detected to be initialised which allows the bootstrap container (or
any container in a pod it has mounted that path into) update the
hostname at bootstrap time.

Bootstrap has been updated to write to this file.

NB: This file is only read once - once kubelet has been restarted it is
no longer possible to rename the host.

This will cause a problem for reboots, as the bootstrap code does not
currently run for non fresh boots - I think I will be changing this
soon, as there are some instances where an every-time init is more
appropriate.