Fix the init binary copy
The init exe is unique in that it sits at /init (rather than /bin/init
like everything else).

The build-pkg is not setup to deal with this use case and instead tries
to standardise all binaries into the /bin directory (this is correct
behaviour!).

So, rather than labelling /init as an initrd_binary, we're now just
manually copying it into place.

This does technically mean that build-pkg won't run its lib check on it,
however init does not depend on anything other than musl, which is
guaranteed to exist, so this is not an issue.