Force chrony to build without libcap
We do not want chrony to be worrying about trying to set its own
capabilities (as that is handled for us as part of the build / container
securityContext.

Previously the configure command was occassionally picking up that
libcap was present (as it would be in the build directory if crun
happened to have been built first). This sets configure to always ignore
libcap, even if it detects it.