Add patch to revert bridge shutdown behaviour
[This change][upstream] originally caused the bridge to clean up the
interface before calling the DHCP plugin to clean up.

Unfortuneately, this caused an error as DHCP requires the interface to
still exist so it can release the IP address to the DHCP server.

This patch reverts it as it is required in kios. Please see [the
upstream issue][issue] for further discussion on this.

[upstream]: https://github.com/containernetworking/plugins/pull/702
[issue]: https://github.com/containernetworking/plugins/issues/791