Correctly remove ash from busybox
Although we had disabled ash in busybox's config, we had left the
"default shell" setting on ash, which was causing the ash feature to be
automatically re-enabled at build time.

This corrects that misconfiguration - ash is now excluded, which has cut
down the size of the busybox binary by ~78k (and has vastly improved the
security of the host system, as it no longer possesses the ability to
run arbitrary shell commands.