Remove portmap plugin from DHCP-CNI image
This plugin was not used by the DHCP-CNI plugin and was only included as
a means of auto-installing portmap for k8s hostPort compatibility. It
was not an appropriate way to install this plugin, which is unrelated to
this container image.

In addition, it turns out that cri-o hard codes the portmap (and lo)
logic, so neither of these plugins are required on kiOS - they will
therefore not be included as part of the base system.