Move hostname set to before kubelet start
Previously, we'd check if the initial kubelet run had resulted in a pod
creating a "hostname" file - if it had we would update the hostname.

Now that we include a disk-backed etc directory, it is reasonable to
check for a hostname file _before_ attempting to run any of the
container runtime.

In addition, if a pod needs to change the hostname, it should now do it
via assuming the host's UTS namespace.