Skip pre kubelet init if config file exists
Previously, it was essential that the init kubelet would run on every
boot, as it had to setup files in the ephemeral etc directory. This is
no longer a requirement as etc is disk-backed. This means it is
acceptable to skip the init kubelet setup if the config file already
exists.