Remove pidfiles from chronyd
Pidfiles were being stored in an emptyDir to avoid giving chrony write
access to its root file system. However this meant that the pidfile that
was written here would persist between reboots if there was a crash.
This was preventing chrony from starting in most cases.

The best fix for this is to remove the functionality of using a pidfile.

Fixes #12