Readd support for post-bootstrap hostname setting
We used to support setting the hostname after kubelet's bootstrap run.
This was removed in 377108fb76f965d78d4b8125056565b83a94dfb9 in favour
of the more "correct" approach of having privileged pods assume the
host's UTS namespace and setting the hostname directly.

However, assuming the host's UTS from within a container is a bit
clunky, and the hostname is set by the provider for most cloud setups,
so it is likely that a lot of bootstrap pods are going to need to do
this.

It makes more sense, therefore, to offer hostname setting support as a
convenience here.