Fix directory timing and permissions
Creating the directory after starting the kubelet sometimes leads to a
race condition of non matching inodes between the version kubelet has
and the one init is watching. Also directories were missing the x
permission. While this had not been noticed, it caused some problems for
the bootstrap pod when saving files.