Add support for proper shutdown handling
This adds support for building the "tiny power button" kernel module,
which will listen for Power Button pushes and will automatically signal
init.

As a result signal based system shutdowns have been implemented in init.
This is useful for backwards compatibility also, however the
_recommended_ way for container-space processes to signal for a shutdown
remains to be to use the system sock, as access can be granted to this
without full full hostPID.

The exact process involved in requesting that this module be loaded is
not yet determined. It is most likely that it will be the job of a
bootstrap pod to handle this.

This fixes #9.