Add support for setting kubelet node labels
Kubelet supports setting node labels on intial registration. However,
these can only be passed to it via the command line, unlike most other
settings, which can be given via the standard KubeletConfiguration file.
To ensure that this feature is still supported on kiOS, we have added a
`node-labels` file, which has a list of key: value pairs (one per line)
which are read by kiOS init, rewritten as command line flags, and then
passed to the kubelet when starting it.

Fixes #11