Add kmod
This is predominantly intended to be used as an initContainer for pods
that know they require specific kernel modules (often this will be early
static / bootstrap pods) although kiOS-core will not be prescriptive
about its usage.

This does completely replace udev, which still needs a longer term
k8s-oriented plan, however it unblocks a number of scenarios that needed
to load named modules for the system to start/function correctly.

Fixes #13