Update kubelet to prevent static pod restarts
Previously, when the system's hostname was updated and then the kubelet
restarted, this would result in all static pods being restarted, due to
the internal change of the pod's name.

This updates the way kubelet handles pods internally, so that hostname
changes no longer result in the pods being restarted on hostname
changes, whilst retaining the kubelet's ability to create a mirror pod
in the api-server with its hostname appended.

Fixes #15