Bind /etc/seccomp to /var/lib/kubelet/seccomp
This directory is not configurable in the kubelet, however it makes more
sense for this to exist in the /etc/ directory (so that changes can
exist on disk).

So we will bind mount /etc/seccomp to this directory. We use bind
mounting rather than a symlink / patching kubelet so that any containers
that mount the seccom directory hostPath in the expectation that it will
exist in the "standard" location can find it.