Remove support for automatic kubelet restart
Previously, if the system started without a kubernetes configuration
file, kiOS would start kubelet then watch for the configuration file to
appear. Once it did, it would auto restart the kubelet.

While this was a nice "magic" feature, it had a couple of drawbacks:

- It required bootstrap containers to be aware that they need to write
  the config file last - this is not always clear
- It was relatively black and white - auto restarting only works when
  the system has no configuration at all. Partial configuration could
  not use this method.
- Restarting the container runtime could not be done using this method
- The implementation was poor

For the moment, therefore, this functionality is removed - instead
bootstrap containers should explicitly restart the kubelet by requesting
it via the system socket.