Fix configure host / build targets
We were previously setting the host to be GNU based. This is incorrect
as the host is the final kiOS system, which is musl based. The build
machine is now also explictly set to be GNU x86_64*.

This technically means that building kiOS on ARM or Musl based machines
is not currently possible, but we'll cross that bridge if and when it is
ever required.