Explicitly set PREFIX to /
Previously we were using an empty prefix, which in most cases means the
same as /, however some configure scripts were taking an empty string to
mean "use the default PREFIX" which was causing problems. Explicitly
setting this to the non-empty "/" fixes this. For scripts that did not
interpret the empty string this way, the worst case is that they use
some paths with a superfluous extra slash in their build paths
(eg "/path/to/build/dir//lib") however this does not matter.