Update datapart path creation logic
Previously when kiOS autodetects the data partition, it assumes that the
drive will always be nvme. As nvme drives tend to end in a digit, Linux
separates the block name from the part number with a "p", eg
/dev/nvme0n1p2. The kiOS code assumed this would always be the case, so
would erroneously use the "p" for other types of drives, eg
"/dev/sdap2".

This has been updated to check the name of the block device, and only
add a "p" if it already ends in a digit.