Add naive support for waiting for the datapart
When specifying a datapart explicitly via the datapart command line
parameter, kiOS will already wait for the device to appear in order to
mount it. However, until previously, auto discovery would run once at
startup - if the desired devices was not yet ready, it would never be
found.

Auto discovery will now wait if it was not able to find the root device
and will continually retry.