Build glib libraries as static archives
Rather than building glib and pcre as shared libraries, we now build
them as static archives.

This is useful as they are only used by the conmon binary, so they are
now embedded directly into it. Thanks to the link time optimisation, the
conmon build is able to throw away quite a lot of the glib code and only
a small subset is included in the final binary.

Fixes #25