Update cacerts pulling logic
We now have a top level script which saves the etag and uses it to
compare and repull the cacerts when requested. As part of this change,
we also commit the pulled ca-bundle.pem and etag file to file.

This is to make it slightly clear which SSL certificates are going to be
included.

Fixes #16