Remove explicit call to autogen.sh
Previously each component that used the configure build system was
responsible for explicitly calling ./autogen.sh itself, if required. This
led to a lot of code repetition, and resulting in autogen being called
on every invocation, even when not required.

Moving this to the general build_configuration function allows us to
generalise the call away, and only make it when required (when --clean
is passed or the component has not already been built).