Fix kubelet subdirectory creation
We were mistakenly creating /var/lib/kubelet's `pods` and `seccomp`
directories in the root cpio - this is not needed as init overmounts
this whole directory with tmpfs.

Init has been corrected to create the `seccomp` directory along with
`pods`. The file permission on pods has also been updated to drop the
write bit.