Remove call to bring up eth0
This was originally a useful call for kiOS to make to ensure that IP
networking was available as soon as possible. However, for most cases,
the system is not ready for this, which makes the call here pointless:
for IPv4 systems, a user space daemon is required to acquire IP addresses
anyway - typically this is DHCPCD, which will bring the interface up
itself. Even for IPv6 systems, most (all?) Ethernet devices will require
a module to be loaded before networking will work.

Even worse, making this call when eth0 does not exist will cause the
kernel to try a few module names, exec'ing modprobe each time. This is a
waste of time, and not worth doing.

Removing this is ultimately more in line with the kiOS core "way" of
having containers do the setup, not kiOS itself.