Add audit seccomp policy to root image
This bundles a generic seccomp policy, which audits all syscalls, into
the initramfs root. This is a useful generic to have available by
default, so that users can create their own custom policies.

I originally played around with the idea of having a few more "default"
polices bundled in kiOS (eg a "sensible defaults" similar to docker's)
however seccomp is so specific that it is difficult to come up with a
meaningful generic. If a user is opting for seccomp (as opposed to bog
standard capability limiting) it is likely that they already have a
specific policy in mind.

Fixes #22