socket: Update system to be an event loop
Previously, we spawned a dedicated thread for a pre-determined amount of
listeners for the system socket, and at runtime, init kept itself alive
with a waitpid loop to reap zombie children.

We have now dropped the wait loop, and instead moved the system socket
code to be a long running event loop instead. The advantage of this
approach, which uses the kernel's epoll, is that we can add other file
descriptors to the event loop later on as required.

The system sockets were also relatively low usage, so didn't really need
dedicated threads per listener - we can also get away with having more
clients now.