Ensure that tar end earlies are respected
When tar reaches a dead block, it immediately returns, signalling that
it is done. Gzip is coded to honour this and immediately treat this as
the end of the stream, however in the case that deflate still had not
finished decoding that input chunk, it would finish that off first.

While not particularly problematic, this wasted time, and meant tar
could not rely on not being called again after finishing. Which it now
can.

In addition, we have amended the tar callback to properly pass the
status to gzip.