Remove explicit root flag from cri-o
The previous thinking was very much that kiOS should enforce that crio
works in a particular way, as anything else is unsupported / won't work.
However, as it turns out it is relatively easy for a bootstrap container
to manage the datapart.

In order for this flow to work, I have found that giving the bootstrap
container the ability to modify these more advanced settings can be
beneficial.