Switch to always using KubeletConfiguration
Previously, kiOS would check for the presense of a config.yaml file. If
found, it would pass this in when calling `kubelet`. If not, it would
pass in a default static pod path. It would also always pass in the
container runtime path.

This behaviour is a hangover from kiOS' old behaviour - which was to
wait for the existance of the config file to restart the kubelet.

Kubelet is becoming more explictly opposed to command line flags, and
although there is no clear usecase for overriding the container
endpoint, it also adds nothing to be so agressively opinionated about
it. Instead we now include a default config.yaml in the inbuilt root
filesystem, with our defaults, and that will always be used.

This can, and should, be overwritten as appropriate, either during
static provisioning or a bootstrap process.