Instruct kubelet to prefer IPv6
Normally kubelet is able to determine its IP based on the host network.
In the case that there is more than one, it'll need to pick one to be
the "primary IP". For dual stack setups, kubelet normally prefers IPv4,
so will pick that as the primary over any IPv6 addresses that may exist.

This behaviour can be overriden by passing `--node-ip ::` which tells
kubelet to prefer IPv6. If no IPv6 address exists, this option does not
prevent it from picking an IPv4 address, so this is safe to hardcode.