Do not repull images when they already exist
Previously we would always pull the list of containers that we wanted.
However this adds time, and we would rather this utility be something we
can quickly run to confirm the repo is in a state it should be. So
instead we now check if the images exist already and only if they do not
will we attempt to pull them.

This will allow us to pop this command into a number of places without
adding lots of time. We will add a force / always pull flag in at some
point also.