Chrony: Build via Containerfile
Most of the images build using a weird hybrid system where we first
build them outside of our container runtime, then we use a Containerfile
last minute to copy the correct files in place.

This is clunky and weird - we've moved chrony to simply build everything
inside its Containerfile in the usual way. Additionally it now builds a
statically linked binary, rather than one linked to musl libc, which was
also weird inside a container.