Remove wait-for-time image
This was a previous weird hack that was in place to fix the broken time
when booting from a raspberry pi. We have not used this in a long time
and this area of the codebase has not been maintained. If this is ever
needed again, we will spin it up as a seperate project in line with the
new structure.